ab_contract_playground::fungible_ffi

Module balance

Structs

FungibleBalanceArgs

Data structure containing expected input for external method invocation, eventually calling fungible_balance() on the other side by the host.

InternalArgs

Data structure containing expected input to fungible_balance(), it is used internally by the contract, there should be no need to construct it explicitly except maybe in contract’s own tests

Constants

METADATA

Method metadata, see ContractMetadataKind for encoding details

Functions

fungible_balance⚠

FFI interface into a method, called by the host.